backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib
python-multipart
python-dotenv
email-validator
boto3
requests